requirements.txt
boto3
numpy<2
opencv-python
requests
pandas
PyYAML
Pillow
openpyxl
